Initial commit
Define from and to method given two coordinates

Added the price calculation.

Added method to increase the price depending on type of vehicle.

Refactor calculation of the distance in other method

Added greater_max_distance? method to check if the vehicle can make the transport

Added cheaper_transport to get the cheapest option for the distance

Refactor the price calculation, changing the vehicle depending on user preference.